first stab at interval checking

+++ src/main/scala/app/restlike/broadcast/BroadcastFlash.scala
@@ -40,7 +40,7 @@ object BroadcastRequestJson {
 }
 
 //TODO: durationSeconds should be non-optional
-case class BroadcastFlash(messages: List[String], env: String, durationSeconds: Option[Long])
+case class BroadcastFlash(messages: List[String], env: String, durationSeconds: Option[Int])
 
 case object BroadcastMessage {
   def send(req: Req) = {
